requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
neo4j==5.14.0
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pandas==2.1.3
loguru==0.7.2
python-dotenv==1.0.0
httpx==0.25.1
asyncpg==0.28.0
email-validator==2.1.0 
pyjwt==2.10.1
